move GHOST reader to own plugin submodule
