+++ tests/plugins/ghost/test_reader.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from pyaerocom.io.read_ghost import ReadGhost
+from pyaerocom.plugins.ghost.reader import ReadGhost
 
 
 @pytest.fixture(scope="module")
